requirements.txt
streamlit==1.29.0
faker==21.0.0
fpdf2==2.7.6